+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
     "key-manager",
     "nexus",
     "nexus/authz-macros",
+    "nexus/blueprint-execution",
     "nexus/db-macros",
     "nexus/db-model",
     "nexus/db-queries",
@@ -113,6 +114,7 @@ default-members = [
     "key-manager",
     "nexus",
     "nexus/authz-macros",
+    "nexus/blueprint-execution",
     "nexus/db-macros",
     "nexus/db-model",
     "nexus/db-queries",
@@ -246,6 +248,7 @@ mime_guess = "2.0.4"
 mockall = "0.12"
 newtype_derive = "0.1.6"
 mg-admin-client = { path = "clients/mg-admin-client" }
+nexus-blueprint-execution = { path = "nexus/blueprint-execution" }
 nexus-client = { path = "clients/nexus-client" }
 nexus-db-model = { path = "nexus/db-model" }
 nexus-db-queries = { path = "nexus/db-queries" }
