+++ nexus/blueprint-execution/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "nexus-blueprint-execution"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+futures.workspace = true
+nexus-db-queries.workspace = true
+nexus-types.workspace = true
+reqwest.workspace = true
+sled-agent-client.workspace = true
+slog.workspace = true
+uuid.workspace = true
+
+omicron-workspace-hack.workspace = true
